move secrets to basic
